# server/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/task/refreshTask": {
            "post": {
                "operationId": "task.refreshTask",
                "summary": "\u91cd\u65b0\u5165\u961f\uff1a\u7b49\u5f85/\u89e3\u6790/\u4f20\u8f93\u4e2d\u5361\u4f4f\u7684\u4efb\u52a1\u91cd\u65b0\u6392\u961f\uff08\u6210\u529f/\u5931\u8d25\u8d70\u4e13\u95e8\u7aef\u70b9\uff09",
                "tags": [
                    "Task"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 200
                                                },
                                                "msg": {
                                                    "type": "null"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 500
                                                },
                                                "msg": {
                                                    "type": "string",
                                                    "const": "\u5931\u8d25\u4efb\u52a1\u8bf7\u4f7f\u7528\u91cd\u8bd5"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 500
                                                },
                                                "msg": {
                                                    "type": "string",
                                                    "const": "\u5df2\u5b8c\u6210\u7684\u4efb\u52a1\u65e0\u9700\u91cd\u65b0\u5165\u961f"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 500
                                                },
                                                "msg": {
                                                    "type": "string",
                                                    "const": "\u4efb\u52a1\u4e0d\u5b58\u5728"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/task/errorTaskRetry": {
            "post": {
                "operationId": "task.errorTaskRetry",
                "summary": "\u91cd\u8bd5\u5931\u8d25\u4efb\u52a1",
                "tags": [
                    "Task"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 200
                                                },
                                                "msg": {
                                                    "type": "null"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 500
                                                },
                                                "msg": {
                                                    "type": "string",
                                                    "const": "\u4ec5\u5931\u8d25\u4efb\u52a1\u53ef\u91cd\u8bd5"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 500
                                                },
                                                "msg": {
                                                    "type": "string",
                                                    "const": "\u4efb\u52a1\u4e0d\u5b58\u5728"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            }
        }
    }
}